+++ R/messages.R
@@ -59,13 +59,13 @@ info <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path")
 ) {
-  (is_scalar_integerish(level) && level > 0 && level <= 10) ||
+  (rlang::is_scalar_integerish(level) && level > 0 && level <= 10) ||
     stop("'level' must be an integer between 1 and 10: ", level)
-  (is_scalar_integerish(msg_level) && msg_level > 0 && msg_level <= 10) ||
+  (rlang::is_scalar_integerish(msg_level) && msg_level > 0 && msg_level <= 10) ||
     stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
   (
     (is_null(msg_types) || is_character(msg_types)) &&
-      is_in(msg_types, c("INFO", "WARNING", "ERROR"))
+      all(is_in(msg_types, c("INFO", "WARNING", "ERROR")))
   ) ||
     stop(
       "'msg_types' must be NULL (no messages) or a character vector containing",
@@ -162,13 +162,13 @@ warn <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path")
 ) {
-  (is_scalar_integerish(level) && level > 0 && level <= 10) ||
+  (rlang::is_scalar_integerish(level) && level > 0 && level <= 10) ||
     stop("'level' must be an integer between 1 and 10: ", level)
-  (is_scalar_integerish(msg_level) && msg_level > 0 && msg_level <= 10) ||
+  (rlang::is_scalar_integerish(msg_level) && msg_level > 0 && msg_level <= 10) ||
     stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
   (
     (is_null(msg_types) || is_character(msg_types)) &&
-      is_in(msg_types, c("INFO", "WARNING", "ERROR"))
+      all(is_in(msg_types, c("INFO", "WARNING", "ERROR")))
   ) ||
     stop(
       "'msg_types' must be NULL (no messages) or a character vector containing",
@@ -268,13 +268,13 @@ error <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path")
 ) {
-  (is_scalar_integerish(level) && level > 0 && level <= 10) ||
+  (rlang::is_scalar_integerish(level) && level > 0 && level <= 10) ||
     stop("'level' must be an integer between 1 and 10: ", level)
-  (is_scalar_integerish(msg_level) && msg_level > 0 && msg_level <= 10) ||
+  (rlang::is_scalar_integerish(msg_level) && msg_level > 0 && msg_level <= 10) ||
     stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
   (
     (is_null(msg_types) || is_character(msg_types)) &&
-      is_in(msg_types, c("INFO", "WARNING", "ERROR"))
+      all(is_in(msg_types, c("INFO", "WARNING", "ERROR")))
   ) ||
     stop(
       "'msg_types' must be NULL (no messages) or a character vector containing",
@@ -359,13 +359,13 @@ info_if <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path")
 ) {
-  (is_scalar_integerish(level) && level > 0 && level <= 10) ||
+  (rlang::is_scalar_integerish(level) && level > 0 && level <= 10) ||
     stop("'level' must be an integer between 1 and 10: ", level)
-  (is_scalar_integerish(msg_level) && msg_level > 0 && msg_level <= 10) ||
+  (rlang::is_scalar_integerish(msg_level) && msg_level > 0 && msg_level <= 10) ||
     stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
   (
     (is_null(msg_types) || is_character(msg_types)) &&
-      is_in(msg_types, c("INFO", "WARNING", "ERROR"))
+      all(is_in(msg_types, c("INFO", "WARNING", "ERROR")))
   ) ||
     stop(
       "'msg_types' must be NULL (no messages) or a character vector containing",
@@ -450,17 +450,17 @@ warn_if <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path")
 ) {
-  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+  (rlang::is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
     error("'level' must be an integer between 1 and 10: ", level)
   (
-    is_scalar_integerish(msg_level) &&
+    rlang::is_scalar_integerish(msg_level) &&
       isTRUE(msg_level > 0) &&
       isTRUE(msg_level <= 10)
   ) ||
     error("'msg_level' must be an integer between 1 and 10: ", msg_level)
   (
     (is_null(msg_types) || is_character(msg_types)) &&
-      is_in(msg_types, c("INFO", "WARNING", "ERROR"))
+      all(is_in(msg_types, c("INFO", "WARNING", "ERROR")))
   ) ||
     error(
       "'msg_types' must be NULL (no messages) or a character vector containing",
@@ -544,17 +544,17 @@ error_if <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path")
 ) {
-  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+  (rlang::is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
     error("'level' must be an integer between 1 and 10: ", level)
   (
-    is_scalar_integerish(msg_level) &&
+    rlang::is_scalar_integerish(msg_level) &&
       isTRUE(msg_level > 0) &&
       isTRUE(msg_level <= 10)
   ) ||
     error("'msg_level' must be an integer between 1 and 10: ", msg_level)
   (
     (is_null(msg_types) || is_character(msg_types)) &&
-      is_in(msg_types, c("INFO", "WARNING", "ERROR"))
+      all(is_in(msg_types, c("INFO", "WARNING", "ERROR")))
   ) ||
     error(
       "'msg_types' must be NULL (no messages) or a character vector containing",
@@ -653,7 +653,7 @@ assert <- function(
     error("'msg_level' must be an integer between 1 and 10: ", msg_level)
   (
     (is_null(msg_types) || is_character(msg_types)) &&
-      is_in(msg_types, c("INFO", "WARNING", "ERROR"))
+      all(is_in(msg_types, c("INFO", "WARNING", "ERROR")))
   ) ||
     error(
       "'msg_types' must be NULL (no messages) or a character vector containing",