+++ R/msgr-package.R
@@ -19,7 +19,8 @@
 #'
 #' @author Chad Goymer \email{chad.goymer@@gmail.com}
 #'
-#' @importFrom rlang is_integerish is_scalar_integerish
 #' @import purrr
+#'
+#' @noRd
 #'
 globalVariables(c(".", ".data"))